On the way

+++ controller/userController.js
@@ -96,9 +96,22 @@ const allUsers = async (req, res, next) => {
 
 const deleteUser = async (req, res, next) => {
   try {
-    const user = req.user;
+    const targetedUser = req.params.id || req.user._id;
+
+    const user = await User.findById(targetedUser);
+    if (!user) {
+      return next(new HttpError("user not found", 404));
+    }
+    if (
+      req.user.role !== "admin" &&
+      req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
+    }
     await User.deleteOne(user);
-    await cloudinary.uploader.destroy(user.cloudinaryId);
+    if (user.cloudinaryId) {
+      await cloudinary.uploader.destroy(user.cloudinaryId);
+    }
     res
       .status(200)
       .json({ success: true, message: "user deleted successfully" });
@@ -109,24 +122,36 @@ const deleteUser = async (req, res, next) => {
 
 const update = async (req, res, next) => {
   try {
-    const user = req.user;
+    const targetedUser = req.user._id || req.params.id;
+    const user = await User.findById(targetedUser);
     if (!user) {
       return next(new HttpError("user not found", 404));
     }
     const updates = Object.keys(req.body);
 
     const allowedFields = ["name", "password", "phone", "profilePic"];
+    if (req.user.role === "admin") {
+      allowedFields = [...allowedFields, "role", "isVerified"];
+    }
 
     const isValid = updates.every((field) => allowedFields.includes(field));
 
     if (!isValid) {
       return next(new HttpError("only allowed field can be updated", 400));
     }
+    if (
+      req.user.role !== "admin" &&
+      req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
+    }
 
     updates.forEach((update) => (user[update] = req.body[update]));
 
     if (req.file) {
-      await cloudinary.uploader.destroy(user.cloudinaryId);
+      if (user.cloudinaryId) {
+        await cloudinary.uploader.destroy(user.cloudinaryId);
+      }
       user.profilePic = req.file.path;
       user.cloudinaryId = req.file.filename;
     }
